+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+<<<<<<< HEAD
 - [ENH] Add `dropna` parameter to groupby_agg. @samukweku
+=======
+- [ENH] `complete` adds a `by` parameter to expose explicit missing values per group, via groupby. @samukweku
+
+>>>>>>> 27afa39096db4ac83dd045a93a026b8229a754d4
 ## [v0.20.13] - 2021-02-25
 
 -   [ENH] Performance improvements to `expand_grid`. @samukweku